Completed the requirements for the Middleware REST Methods and Postman lesson while following along with the directions, the Node course on video by FreeCodeCamp, using GitHub Copilot for code completion and debugging, and getting some help from CTD mentor Meet Karnik during a half-an-hour long, one-on-one video conference last Friday. Used instructor provided code as well.
